LPS-53904 Remove ClusterBase.
