+++ portal-impl/src/com/liferay/portal/cluster/ClusterExecutorImpl.java
@@ -70,7 +70,6 @@ import java.util.concurrent.ExecutorService;
  */
 @DoPrivileged
 public class ClusterExecutorImpl
-	extends ClusterBase
 	implements ClusterExecutor, PortalInetSocketAddressEventListener {
 
 	@Override
@@ -226,6 +225,11 @@ public class ClusterExecutorImpl
 		}
 
 		return _clusterNodeAddresses.containsKey(clusterNodeId);
+	}
+
+	@Override
+	public boolean isEnabled() {
+		return PropsValues.CLUSTER_LINK_ENABLED;
 	}
 
 	@Override
