+++ portal-impl/src/com/liferay/portal/cluster/JGroupsClusterChannel.java
@@ -179,6 +179,8 @@ public class JGroupsClusterChannel implements ClusterChannel {
 			return;
 		}
 
+		_initSystemProperties();
+
 		String autodetectAddress = PropsValues.CLUSTER_LINK_AUTODETECT_ADDRESS;
 
 		if (Validator.isNull(autodetectAddress)) {
@@ -228,6 +230,29 @@ public class JGroupsClusterChannel implements ClusterChannel {
 		}
 
 		_initBindAddress = true;
+	}
+
+	private void _initSystemProperties() {
+		for (String systemProperty :
+				PropsValues.CLUSTER_LINK_CHANNEL_SYSTEM_PROPERTIES) {
+
+			int index = systemProperty.indexOf(CharPool.COLON);
+
+			if (index == -1) {
+				continue;
+			}
+
+			String key = systemProperty.substring(0, index);
+			String value = systemProperty.substring(index + 1);
+
+			System.setProperty(key, value);
+
+			if (_log.isDebugEnabled()) {
+				_log.debug(
+					"Setting system property {key=" + key + ", value=" + value +
+						"}");
+			}
+		}
 	}
 
 	private static final Log _log = LogFactoryUtil.getLog(
